requirements.txt
Django==4.2.10
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.2.0
opencv-python==4.9.0.80
pytrends==4.9.2
tweepy==4.14.0
praw==7.7.1
newsapi-python==0.2.7
openai==1.12.0
diffusers==0.26.3
transformers==4.37.2
torch==2.2.0
boto3==1.34.34
python-instagram==0.8.0
gunicorn==21.2.0
requests==2.31.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0
whitenoise==6.6.0
beautifulsoup4==4.12.3
lxml==5.1.0 